cpp.util: tests and bench work
---AI generated commit message:
Refactor and enhance utility classes in the VOLTISO library
- Updated SplayTable to improve accessors and streamline existence checks.
- Refactored Storage class to support relocation and improve memory management.
- Modified Subject class to encapsulate value handling and improve access patterns.
- Enhanced Array and DynamicArray classes for better item access and memory safety.
- Updated DynamicString and String classes to use new casting methods.
- Improved test cases for BitArray, Handle, Owned, SplayMap, SplaySet, and Storage to reflect changes in object handling and ensure correctness.
- Introduced a new relocate utility to facilitate safe object relocation for trivially relocatable types.

+++ packages/cpp/util/bench/Owned.bench.cpp
@@ -25,10 +25,10 @@ struct LargeObject final {
   std::vector<int> data;
   std::string name;
 
-  explicit LargeObject(size_t size = 100, const std::string &n = "test")
-      : data(size, 1), name(n) {}
+	explicit LargeObject(int size = 100, const std::string &n = "test")
+	    : data(size, 1), name(n) {}
 
-  int sum() const {
+	int sum() const {
     int result = 0;
     for (auto val : data)
       result += val;
